Added Code
Converted "experience" to integer values. for some reason it breaks if I already use the integered experience from tidy_data? It works now though. (don't touch it if it works) See the code block(s) for more notes.

Added the code for a standard knn regression model. for some reason it doesn't like to bind_cols at the end? Someone should take a look at that.